requirements.txt
python-telegram-bot==20.7
asyncio==3.4.3
aiosqlite==0.19.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
schedule==1.2.0
cryptography==46.0.0
Pillow==10.1.0